streamlit_dashboard/requirements.txt
streamlit
pandas
psycopg2-binary
streamlit-autorefresh
docker